crewai-readme-gen/requirements.txt
python-dotenv
crewai
crewai_tools
anthropic
langchain-anthropic
langchain-community
pathspec
duckduckgo-search
